requirements.txt
git+https://github.com/ape364/aiobittrex.git
aiodns==2.0.0
aiogram==2.0.1
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
Babel==2.6.0
beautifulsoup4==4.7.1
cchardet==2.1.4
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
idna==2.8
multidict==4.5.2
peony-twitter==1.1.2
pycares==3.0.0
pycparser==2.19
Pympler==0.6
python-dotenv==0.10.1
pytz==2018.9
PyYAML==5.1
soupsieve==1.9
ujson==1.35
websockets==7.0
yarl==1.3.0
